Temp Fix for Repeat Build Order
I created a custom function that will run in an infinite loop, it will check if the requested Factory is Idle before it constructs these units. This is only a Temp Fix, it has been tested and it works. However, the result of whether the unit is still produced when the player requests a unit is unknown.